requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.0.0
pyyaml>=6.0.0
feedparser>=6.0.0
